# swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: "DISIer"
  description: "E' possibile fare delle richieste a DISIer attraverso delle API. Queste API permettono di interagire con il bot e di scoprire se una certa aula è libera in quel momento."
  termsOfService: http://swagger.io/terms/
  contact:
    name: DISIer Team
    email: davide.farina-1@studenti.unitn.it
    url: https://is2-progetto.herokuapp.com
  license:
    name: MIT
    url: http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT
host: is2-progetto.herokuapp.com
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /bot_message:
    get:
      description: |
        Risponde alla domanda fatta dall'utente. Al bot si possono chiedere informazioni riguardo agli orari degli autobus.
      operationId: botMessageGet
      parameters:
        - name: msg
          in: query
          description: Domanda da porre al bot
          required: true
          type: string
      responses:
        200:
          description: Risposta del bot
          schema:
            type: array
            items:
              $ref: '#/definitions/BotResponse'
        default:
          description: Errore
          schema:
            $ref: '#/definitions/Error'
  /aule:
    get:
      description: |
        Fornisce gli eventi in corso in ogni aula delle sedi di Povo per ogni ora del giorno corrente.
      responses:
        200:
          description: Risposta da aule
          schema:
            type: array
            items:
              $ref: '#/definitions/jsonAule'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /orari:
    get:
      description: |
        Fornisce gli orari del corso di studi nella settimana corrente.
      parameters:
        - name: corso
          in: query
          description: Nome del corso di studi
          required: true
          type: string
        - name: anno2
          in: query
          description: Anno relativo al corso di studi
          required: true
          type: string
      responses:
        200:
          description: Risposta da orari
          schema:
            type: array
            items:
              $ref: '#/definitions/jsonOrari'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  BotResponse:
    required:
      - speech
      - displayText
      - data
      - contextOut
      - source
    properties:
      speech:
        type: string
      displayText:
        type: string
      data: 
        type: object
      contextOut:
        type: array
      source:
        type: string
  jsonAule:
    required:
        - fasce
        - config
        - area_rooms
        - contains_events
        - n_events
        - events
        - rooms
        - legenda
        - table
        - order
        - colori
    properties:
        fasce:
          type: array
        config: 
          type: object
        area_rooms: 
          type: object
        contains_events:
          type: integer
          format: int32
        n_events:
          type: integer
          format: int32
        events:
          type: array
        rooms: 
          type: object
        legenda:
          type: array
        table: 
          type: object
        order:
          type: array
        colori:
          type: array
  jsonOrari:
    required:
        - day
        - giorni
        - fasce
        - granularita
        - start_slot
        - end_slot
        - colori
        - first_day
        - last_day
        - anno
        - lang
        - num_fasce
        - fisrt_day_label
        - last_day_label
        - tipo
        - cds
        - anno2
        - celle
        - legenda
        - note_generali
        - date_inizio_fine_curriculum
        - aa
        - all_events
    properties:
        day:
          type: string
        giorni:
          type: array
        fasce:
          type: array
        granularita:
          type: string
        start_slot:
          type: string
        end_slot:
          type: string
        colori:
          type: array
        first_day:
          type: string
        last_day:
          type: string
        anno:
          type: string
        lang:
          type: string
        num_fasce:
          type: integer
          format: int32
        fisrt_day_label:
          type: string
        last_day_label:
          type: string
        tipo:
          type: string
        cds:
          type: string
        anno2:
          type: string
        celle:
          type: array
        legenda:
          type: array
        note_generali:
          type: array
        date_inizio_fine_curriculum:
          type: array
        aa:
          type: string
        all_events:
          type: integer
          format: int32
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string